Revert changes in normalize method because it affected ajax validation
Ajax validation uses Html::getInputId to generate the input ID, which in turn converts the string to lower case. This is why if we remove the same behavior from BaseColumn::normalize, the ajax validation will not work because the input (attribute) IDs from the ajax response will not match the form input IDs generated by our widget.